feat: arranged message folders
