+++ src/config/messages.ts
@@ -1,0 +1,3 @@
+export const configMessage = {
+  errorFetchingActiveProjectId: 'Error while fetching the active project id.',
+}
