+++ src/extension.ts
@@ -35,7 +35,7 @@ import { telemetryHelpers } from './features/telemetry/telemetryHelpers'
 import { verifyVC } from './features/verify/verifyVC'
 import { initIam } from './features/iam/initIam'
 import { notifyError } from './utils/notifyError'
-import { schemaMessage } from './messages/messages'
+import { schemaMessage } from './features/schema-manager/messages'
 
 const GITHUB_ISSUES_URL = 'https://github.com/affinidi/vscode-extension/issues'
 const GITHUB_NEW_ISSUE_URL = 'https://github.com/affinidi/vscode-extension/issues/new'
