+++ src/features/issuance/csvCreationService.ts
@@ -5,8 +5,7 @@ import { showQuickPick } from '../../utils/showQuickPick'
 import { parseUploadError } from './csvUploadError'
 import { issuanceClient } from './issuanceClient'
 import { ext } from '../../extensionVariables'
-import { csvMessage, snippetMessage, issuanceMessage } from './messages'
-import { labels } from '../../messages/messages'
+import { csvMessage, snippetMessage, issuanceMessage, labels } from './messages'
 import { schemaManagerHelpers } from '../schema-manager/schemaManagerHelpers'
 import { iamState } from '../iam/iamState'
 import { configVault } from '../../config/configVault'
