+++ src/features/messages.ts
@@ -1,0 +1,7 @@
+export const errorMessage = {
+    unexpectedResourceType: 'Unexpected resource type:',
+  }
+  
+  export const labels = {
+    select: 'Select',
+  }
