+++ src/features/schema-manager/schema-builder/helpers/createBuilderSchemaFork.ts
@@ -1,7 +1,7 @@
 import { SchemaField, parseSchema } from '@affinidi/affinidi-vc-schemas';
 import { nanoid } from 'nanoid';
 import { window, ProgressLocation } from 'vscode';
-import { schemaMessage } from '../../../../messages/messages';
+import { schemaMessage } from '../../messages';
 import { vcJsonSchemaFetcher } from '../../../issuance/json-schema/json-schema-fetcher';
 import { schemaManagerState } from '../../schemaManagerState';
 import { BuilderAttribute } from '../SchemaBuilderWebview';
