+++ src/features/schema-manager/tree/schemaManagerExplorerProvider.ts
@@ -1,6 +1,5 @@
 import { ThemeIcon } from 'vscode'
-import { labels } from '../../../messages/messages'
-import { schemaMessage } from '../messages'
+import { schemaMessage, labels } from '../messages'
 import { BasicTreeItem } from '../../../tree/basicTreeItem'
 import { ExplorerProvider } from '../../../tree/explorerTree'
 import { Feature } from '../../feature'
