+++ src/features/telemetry/messages.ts
@@ -1,0 +1,11 @@
+export const telemetryMessage = {
+    accept: 'Accept',
+    deny: 'Deny',
+    sendUsageData: 'Help us make Affinidi better! Do you accept to send anonymous usage data?',
+  }
+
+
+  export const errorMessage = {
+    unknownConsentType: 'unknown consent type: ',
+  }
+
