+++ src/generators/messages.ts
@@ -1,0 +1,12 @@
+export const generatorMessage = {
+    failedToGenerateApp: 'Failed to generate app',
+    noDirectorySelected: "Installation folder wasn't selected",
+    appNameNotSelected: "App name wasn't specified",
+    checkCliInstall: 'Checking CLI installation...',
+    enterAppName: 'Enter an app name',
+    directoryNameDuplication: 'Directory with this name already exist',
+  }
+
+  export const labels = {
+    select: 'Select',
+  }
