+++ src/tree/devToolsTree.ts
@@ -1,5 +1,5 @@
 import { ThemeIcon, TreeDataProvider, TreeItem, TreeItemCollapsibleState } from 'vscode'
-import { labels } from '../messages/messages'
+import { labels } from './messages'
 import { DevToolsTreeItem } from './devToolsTreeItem'
 
 export enum DevToolsType {
