+++ src/tree/messages.ts
@@ -1,0 +1,21 @@
+export const labels = {
+    schemaBuilder: 'Schema Builder',
+    initiateIssuanceCsvFlow: 'Initiate CSV Issuance',
+    select: 'Select',
+    empty: '(empty)',
+    viewAvailableSnippets: 'View Available Snippets',
+    appGenerators: 'App Generators',
+    scriptGenerators: 'Script Generators',
+    intelliSenseSnippets: 'IntelliSense Snippets',
+    helperTools: 'Helper Tools',
+    sendVCOfferToEmail: 'Send a VC Offer to an email',
+    getIssuanceOffers: 'Get Issuance Offers',
+    signCloudWalletVc: 'Sign a VC with Cloud Wallet',
+    certificationAndVerification: 'Certification & Verification',
+    reviewIssues: 'Review Issues',
+    reportIssue: 'Report Issue',
+    openWalkthrough: 'Extension Walkthrough',
+    openAPIDocs: 'API Documentation',
+    openDiscord: 'Discord Server',
+    verifyVC: 'Verify a VC',
+  }
